Code refactoring and bug fixes: - Refactored the IsometricPath class to extend from an abstract class - Fixed a bug that was not returning the IsometricPath class instance for the clear and update methods
