+++ src/@classes/abstract/Graphic/Graphic.ts
