+++ src/@classes/abstract/Graphic/index.ts
