+++ src/@classes/abstract/Graphic/types.ts
