+++ src/@classes/abstract/IsometricGraphic/index.ts
@@ -1,0 +1,1 @@
+export { IsometricGraphic } from './IsometricGraphic';
