+++ src/@classes/public/IsometricCanvas/types.ts
