+++ src/@classes/public/IsometricPath/index.ts
