+++ src/@classes/public/IsometricPath/types.ts
